+++ dejafu/CHANGELOG.rst
@@ -7,6 +7,16 @@ standard Haskell versioning scheme.
 .. _PVP: https://pvp.haskell.org/
 
 
+unreleased
+----------
+
+Added
+~~~~~
+
+* A ``Test.DejaFu.successful`` predicate, to check that a computation
+  never fails.
+
+
 1.9.0.0 (2018-06-10)
 --------------------